+++ src/main/java/com/heymoose/domain/affiliate/counter/BufferedShows.java
@@ -24,7 +24,7 @@ public class BufferedShows extends BufferedCounters {
   public void flushCounter(long key, int diff) {
     sessionProvider.get()
         .createQuery("update OfferStat set showCount = showCount + :diff where id = :id")
-        .setParameter("diff", diff)
+        .setParameter("diff", (long) diff)
         .setParameter("id", key)
         .executeUpdate();
   }
